backend/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
Flask>=3.0.0
flask-cors>=4.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
PyYAML>=6.0
requests>=2.28.0
ultralytics>=8.0.0
pymongo>=4.0.0
gunicorn
